move new feat into original file
